some more actions moved from Swift to Kotlin

+++ iosApp/iosApp/SongListView.swift
@@ -64,7 +64,7 @@ struct SongListView: View {
                                         .highPriorityGesture(
                                             TapGesture()
                                                 .onEnded { _ in
-                                                    self.onPerformAction(SongClick(songIndex: index))
+                                                    self.onPerformAction(SongClick(songIndex: Int32(index)))
                                                     self.songListComponent?.onSongClicked(position: Int32(index))
                                                 }
                                         )
